refactor: 패키지 변경 및 메서드 위치 일관성 유지
